feat: reuse lote summary component
